Added description
This script can run on any device because it does not query the N-Able services for appliance id or N-Central server.  It shows the use of the unique namepsace so it can be run repeatedly in the same powershell session.  It also demonstrates the use of querying for device info based upon a filter on the N-Central server.